pyproject.toml
[project]
name = "python-fastapi-template"
version = "0.1.0"
description = "A template for a Python FastAPI service with uv & Devcontainer"
authors = [{ name = "Shakil Mosharrof", email = "shakilmrf8@gmail.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "agno>=1.2.6",
  "ai-data-science-team>=0.0.0.9016",
  "chromadb>=1.0.0",
  "commitizen>=4.4.1",
  "duckdb>=1.2.1",
  "fastapi[standard]>=0.115.8",
  "fireducks>=1.2.5",
  "ipykernel>=6.29.5",
  "jupyter>=1.1.1",
  "langchain>=0.3.22",
  "langgraph>=0.3.21",
  "matplotlib>=3.9.4",
  "numpy>=2.0.2",
  "openai>=1.69.0",
  "openai-agents>=0.0.7",
  "pandas>=2.2.3",
  "pandas-stubs>=2.2.2.240807",
  "phi>=0.6.7",
  "phidata>=2.7.10",
  "pre-commit>=4.1.0",
  "pydantic-ai[logfire]>=0.0.46",
  "pydantic-graph>=0.0.52",
  "scikit-learn>=1.6.1",
  "streamlit>=1.44.1",
]


[dependency-groups]
dev = [
  "mypy",
  "pre-commit>=4.1.0",
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
  "ruff>=0.9.9",
  "black",
  "isort",
  "pre-commit",
  "uv",
]

[tool.mypy]
strict = true
pretty = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --cov=app --cov-branch --cov-report=term --cov-report=html --cov-fail-under=100"
testpaths = [
    "tests",
]
# You can add other pytest options here if needed
# e.g., filter warnings, etc.

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["app"]
